requirements.txt
flask==2.3.2
gunicorn==20.1.0
python-telegram-bot==20.3
requests==2.31.0
pymongo==4.3.3
python-dotenv==1.0.0 